feat benchmark: OTP app with integrated counter
Turns the benchmark into a constantly running OTP app that prints a req/sec average every 5 seconds.
