+++ lib/riemann_benchmark_elixir.ex
@@ -5,6 +5,8 @@ defmodule RiemannBenchmarkElixir do
     import Supervisor.Spec, warn: false
 
     children = [
+      worker(RiemannBenchmarkElixir.Worker, [1_000_000]),
+      worker(RiemannBenchmarkElixir.Counter, []),
     ]
 
     opts = [strategy: :one_for_one, name: RiemannBenchmark.Supervisor]
